requirements.txt
azure-storage-blob
fastapi
aiohttp